+++ pytorch_lightning/plugins/precision/precision_plugin.py
@@ -1,29 +1,41 @@
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import math
-from typing import Union
+from typing import Generator, Union
 
 import torch
 from torch.optim import Optimizer
 
-from pytorch_lightning.plugins .base_plugin import Plugin
 from pytorch_lightning.core import LightningModule
+from pytorch_lightning.plugins.base_plugin import Plugin
 
 
 class PrecisionPlugin(Plugin):
     EPSILON = 1e-6
     precision = 32
 
-    def pre_optimizer_step(self, optimizer, optimizer_idx):
-        pass
+    def master_params(self, optimizer: torch.optim.Optimizer) -> Generator[torch.Tensor, None, None]:
+        """The master params of the model. Returns the plain model params here.
+        Maybe different in other precision plugins.
 
-    def post_optimizer_step(self, optimizer, optimizer_idx):
-        pass
-
-    def master_params(self, optimizer):
+        """
         for group in optimizer.param_groups:
             for p in group["params"]:
                 yield p
 
     def connect(self, model: torch.nn.Module, optimizers, lr_schedulers):
+        """Connects this plugin to the accelerator and the training process"""
         return model, optimizers, lr_schedulers
 
     def backward(
@@ -36,6 +48,16 @@ class PrecisionPlugin(Plugin):
         *args,
         **kwargs,
     ):
+        """performs the actual backpropagation
+
+        Args:
+            model: the model to be optimized
+            closure_loss: the loss value obtained from the closure
+            optimizer: the optimizer to perform the step lateron
+            opt_idx: the optimizer's index
+            should_accumulate: whether to accumulate gradients or not
+
+        """
         automatic_optimization = model.automatic_optimization
 
         # do backward pass
@@ -50,6 +72,7 @@ class PrecisionPlugin(Plugin):
         return closure_loss
 
     def clip_gradients(self, optimizer: Optimizer, clip_val: Union[int, float], norm_type: float = float(2.0)):
+        """Clips the gradients to a specific value"""
         # TODO: separate TPU case from here
         if clip_val is None:
             return