+++ pytorch_lightning/plugins/training_type/training_type_plugin.py
@@ -1,3 +1,16 @@
+# Copyright The PyTorch Lightning team.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import os
 from abc import ABC, abstractmethod
 from typing import Optional
@@ -5,10 +18,13 @@ from typing import Optional
 import torch
 
 from pytorch_lightning import _logger as log
-from pytorch_lightning.plugins .base_plugin import Plugin
+from pytorch_lightning.core.lightning import LightningModule
+from pytorch_lightning.plugins.base_plugin import Plugin
 
 
 class TrainingTypePlugin(Plugin, ABC):
+    """A Plugin to change the behaviour of the training, validation and test-loop."""
+
     def __init__(self):
         self._model = None
         self._results = None
@@ -16,36 +32,36 @@ class TrainingTypePlugin(Plugin, ABC):
 
     @property
     @abstractmethod
-    def on_gpu(self):
-        raise NotImplementedError
+    def on_gpu(self) -> bool:
+        """Returns whether the current process is done on GPU"""
 
     @property
     @abstractmethod
     def root_device(self) -> torch.device:
-        raise NotImplementedError
+        """Returns the root device"""
 
     @abstractmethod
     def model_to_device(self):
-        raise NotImplementedError
+        """Moves the model to the correct device"""
 
     @property
     @abstractmethod
-    def is_global_zero(self):
-        raise NotImplementedError
+    def is_global_zero(self) -> bool:
+        """Whether the current process is the rank zero process not only on the local node, but for all nodes."""
 
     @abstractmethod
     def reduce(self, output, *args, **kwargs):
-        raise NotImplementedError
+        """Reduces the given output (e.g. across GPUs/Processes)"""
 
     @abstractmethod
     def barrier(self, name: Optional[str] = None):
-        raise NotImplementedError
+        """Forces all possibly joined processes to wait for each other"""
 
     @abstractmethod
     def broadcast(self, obj: object, src: int = 0) -> object:
-        raise NotImplementedError
+        """Broadcasts an object to all processes"""
 
-    # TODO method this is currently unused
+    # TODO method this is currently unused. Check after complete refactors are pushed
     def set_nvidia_flags(self, is_slurm_managing_tasks, device_ids):
         if device_ids is None:
             return
@@ -54,21 +70,24 @@ class TrainingTypePlugin(Plugin, ABC):
         os.environ["CUDA_DEVICE_ORDER"] = "PCI_BUS_ID"
         all_gpu_ids = ",".join([str(x) for x in range(torch.cuda.device_count())])
         devices = os.environ.get("CUDA_VISIBLE_DEVICES", all_gpu_ids)
-        log.info(f'LOCAL_RANK: {self.trainer.local_rank} - CUDA_VISIBLE_DEVICES: [{devices}]')
+        log.info(f"LOCAL_RANK: {self.trainer.local_rank} - CUDA_VISIBLE_DEVICES: [{devices}]")
 
     def reduce_early_stopping_decision(self, should_stop: bool) -> bool:
+        """Reduce the early stopping decision across all possibly spawned processes"""
         return should_stop
 
     @property
-    def model(self):
+    def model(self) -> torch.nn.Module:
+        """Returns the potentially wrapped LightningModule"""
         return self._model
 
     @model.setter
-    def model(self, new_model):
+    def model(self, new_model: torch.nn.Module):
         self._model = new_model
 
     @property
-    def lightning_module(self):
+    def lightning_module(self) -> LightningModule:
+        """Returns the pure LightningModule without potential wrappers"""
         return self._model
 
     @property
@@ -81,13 +100,13 @@ class TrainingTypePlugin(Plugin, ABC):
         return self._results
 
     @property
-    def rpc_enabled(self):
+    def rpc_enabled(self) -> bool:
         return False
 
-    def start_training(self, trainer):
+    def start_training(self, trainer: "Trainer") -> None:
         # double dispatch to initiate the training loop
         self._results = trainer.train()
 
-    def start_testing(self, trainer):
+    def start_testing(self, trainer: "Trainer") -> None:
         # double dispatch to initiate the test loop
         self._results = trainer.run_test()